Changed resx behavior to not rely on resgen tool
So that I can develop in vscode, it was needed to drop
reliance on the resgen tool which gens .designer.cs files
for resx resources. Is a new library now that provides
text via a lookup. Not quite as clean, but hey tell
dotnet core devs to sort resgen out for vscode peeps :-(