Fix setting the right channel on deployments. Switch to Github version checking
If it defaulted to a default channel for a project, OctoPlus never set a channel
id on the release. This made it pick the default for the project, which may not
be the same as the configured default channel.

Moved back to github, so changed the version checking back to use the GitHub
version.